+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
+## Unreleased
+- Added error handling for various loan and grant limits
+
 ## 2.2.2 2016-02-19
 - Further fix to private loan monthly payments and lifetime oan value